test_requirements.txt
# Test requirements for Emphizor integration tests
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
python-dotenv>=1.0.0
freezegun>=1.2.0  # For testing time-dependent functionality
coverage>=7.0.0   # For test coverage reporting

# Include main app requirements
PySide6>=6.0.0
fsrs>=4.0.0
supabase>=2.0.0 